pyproject.toml
[project]
name = "deskmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "fastapi>=0.121.1",
    "google-generativeai>=0.8.5",
    "pydantic>=2.12.4",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
